pyproject.toml
[tool.poetry]
name = "sei-project4-the-community-project-backend"
version = "0.1.0"
description = ""
authors = ["Rachel Tipton <RACHELMTIPTON@GMAIL.COM>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
Pillow = "^9.1.0"
django-summernote = "^0.8.20"
djangorestframework-simplejwt = "^5.1.0"
cors = "^1.0.1"
django-cors-headers = "^3.11.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
